exception

+++ user-service/src/main/java/org/aibles/userservice/service/iml/UserServiceIml.java
@@ -1,10 +1,13 @@
 package org.aibles.userservice.service.iml;
 
 import liquibase.pro.packaged.U;
+import org.aibles.userservice.exception.UserNotFound;
 import org.aibles.userservice.model.User;
 import org.aibles.userservice.repository.UserRepository;
 import org.aibles.userservice.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -26,36 +29,30 @@ public class UserServiceIml implements UserService {
     }
 
     @Override
-    public User getUser(int id) {
+    public User getUser(int id) throws UserNotFound {
         User user = userRepository.findById(id).orElse(null);
-        return user;
-    }
-
-//    @Override
-//    public User updateUser(User user) {
-//        User userUpdated  = userRepository.save(user);
-//        return userUpdated;
-//    }
-
-//    @Override
-//    public User updateUser(int id, User user) {
-//        User userUpdate = userRepository.save(user);
-//        return userUpdate;
-//    }
-
-    @Override
-    public void deleteUser(int id) {
-        userRepository.deleteById(id);
+        if (user == null ) {
+            throw new UserNotFound();
+        }
+        else {
+            return user;
+        }
     }
 
     @Override
-    public void deleteAllUser() {
-        userRepository.deleteAll();
+    public void deleteUser(int id) throws UserNotFound {
+        User user = userRepository.findById(id).orElse(null);
+        if(user == null){
+            throw new UserNotFound();
+        }
+        else {
+            userRepository.deleteById(id);
+        }
     }
 
     @Override
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
+    public User updateUser(int id) throws UserNotFound{
+        return null;
     }
